refactored DeletePlayer to container style
