+++ package.json
@@ -32,5 +32,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "eslint-config-react-app": "^5.0.2",
+    "eslint-config-standard-jsx": "^8.1.0",
+    "eslint-plugin-react": "^7.14.3"
   }
 }
