+++ src/containers/Delete/index.js
@@ -1,0 +1,3 @@
+import Delete from "./Delete";
+
+export default Delete;
